# docs/swagger.yaml
basePath: /
definitions:
...
  doc.Cluster:
    properties:
      active:
        type: boolean
      cloud_storage: {}
      cluster_request:
        $ref: '#/definitions/doc.ClusterRequest'
      cluster_request_id:
        type: integer
      color:
        type: string
      configPath:
        type: string
      context:
        type: string
      createdat:
        type: string
      dns:
        $ref: '#/definitions/doc.DNS'
      dns_id:
        type: integer
      id:
        type: integer
      image_registry:
        $ref: '#/definitions/doc.ImageRegistry'
      image_registry_id:
        type: integer
      labels:
        type: string
      name:
        type: string
      nodes:
        type: integer
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      provider:
        type: string
      provision_percentage:
        type: number
      pv_capacity:
        type: integer
      region:
        type: string
      storage_class:
        type: string
      total_memory:
        type: integer
      weight:
        type: integer
      zone:
        type: string
    type: object
  doc.ClusterRequest:
    properties:
      active:
        type: boolean
      cluster:
        $ref: '#/definitions/doc.Cluster'
      cluster_id:
        type: integer
      cluster_name:
        type: string
      cluster_version:
        type: string
      createdat:
        type: string
      id:
        type: integer
      network_cidr:
        type: string
      network_policy:
        type: boolean
      nfs_detail: {}
      node_group_count:
        type: integer
      node_group_detail: {}
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      provider:
        type: string
      provider_name:
        type: string
      pvc_write_many:
        type: boolean
      region:
        type: string
      regional_cluster:
        type: boolean
      remove_default_node_pool:
        type: boolean
      status:
        type: string
      tls:
        type: string
      type:
        type: string
      vpc_name:
        type: string
      zone: {}
    type: object
...
  doc.DNS:
    properties:
      active:
        type: boolean
      base_domain:
        type: string
      createdat:
        type: string
      id:
        type: integer
      name:
        type: string
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      project_id:
        type: string
      provider:
        type: string
      region:
        type: string
      tls:
        type: string
      zone_id:
        type: string
    type: object
...
  doc.ImageRegistry:
    properties:
      active:
        type: boolean
      createdat:
        type: string
      id:
        type: integer
      name:
        type: string
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      project_name:
        type: string
      provider:
        type: string
      service:
        type: string
      user_name:
        type: string
    type: object
...
  doc.Operator:
    properties:
      active:
        type: boolean
      capabilityLevel:
        type: string
      categories: {}
      channel:
        type: string
      channels: {}
      containerImage:
        type: string
      createdAt:
        type: string
      createdat:
        type: string
      customResourceDefinitions: {}
      description:
        type: string
      displayName:
        type: string
      globalOperator:
        type: boolean
      id:
        type: integer
      k8sMaxVersion:
        type: string
      k8sMinVersion:
        type: string
      keywords: {}
      links: {}
      name:
        type: string
      packageName:
        type: string
      provider:
        type: string
      replaces:
        type: string
      repository:
        type: string
      thumbUrl:
        type: string
      version:
        type: string
      versionForCompare:
        type: string
    type: object
  doc.OperatorRequest:
    properties:
      channel:
        type: string
      cluster:
        $ref: '#/definitions/doc.Cluster'
      cluster_request:
        $ref: '#/definitions/doc.ClusterRequest'
      cluster_request_id:
        type: integer
      createdat:
        type: string
      csv_value:
        type: string
      global_operator:
        type: boolean
      id:
        type: integer
      install_plan_approval:
        type: string
      installation_mode:
        type: string
      operator_details: {}
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      package_name:
        type: string
    type: object
  doc.Organization:
    properties:
      createdat:
        type: string
      description:
        type: string
      domain:
        type: string
      id:
        type: integer
      image:
        type: string
      members:
        items:
          $ref: '#/definitions/doc.OrganizationMembers'
        type: array
      name:
        type: string
      organization_plan:
        $ref: '#/definitions/doc.OrganizationPlan'
      organization_plan_id:
        type: integer
      plugins:
        items:
          $ref: '#/definitions/doc.Plugin'
        type: array
      user:
        $ref: '#/definitions/doc.User'
      user_id:
        type: integer
    type: object
  doc.OrganizationMembers:
    properties:
      createdat:
        type: string
      id:
        type: integer
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      user:
        $ref: '#/definitions/doc.User'
      user_id:
        type: integer
      user_role:
        $ref: '#/definitions/doc.UserRoleEnum'
    type: object
...
  doc.UserRoleEnum:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Admin
    - Members
...
host: api.test.01cloud.dev
info:
  contact:
    email: info@01cloud.com
    name: API Support
  description: 01cloud-api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 01cloud-api
  version: 1.0.1
paths:
...
  /Operators:
    get:
      consumes:
      - application/json
      description: Get list of Operators
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get Operators
      tags:
      - Operators
...
  /admin/operator/{packageName}:
    put:
      consumes:
      - application/json
      description: Update Operator
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update Operator
      tags:
      - Operators
  /admin/operator/enable-disable:
    post:
      consumes:
      - application/json
      description: Enable or Disable Operator
      parameters:
      - description: Package Name
        in: query
        name: packageName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Enable Disable Operator
      tags:
      - Operators
  /admin/operator/sync:
    post:
      consumes:
      - application/json
      description: Sync Operator
      parameters:
      - description: Operator Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.Operator'
      - description: Package Name
        in: query
        name: packageName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Sync Operator
      tags:
      - Operators
  /admin/operator/sync-all:
    post:
      consumes:
      - application/json
      description: Sync All Operator
      parameters:
      - description: Operator
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.Operator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Sync All Operator
      tags:
      - Operators
  /admin/operators:
    get:
      consumes:
      - application/json
      description: Get Operator for Admin
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Operator for Admin
      tags:
      - Operators
...
  /cluster/{id}/operator/{oid}:
    delete:
      consumes:
      - application/json
      description: UnInstall Operator
      parameters:
      - description: Cluster Id
        in: path
        name: id
        required: true
        type: integer
      - description: Operator Id
        in: path
        name: oid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: UnInstall Operator
      tags:
      - Operators
  /cluster/{id}/operator/{oid}/activity-log:
    get:
      consumes:
      - application/json
      description: Get Operator Activity Log
      parameters:
      - description: Cluster Id
        in: path
        name: id
        required: true
        type: integer
      - description: Operator Id
        in: path
        name: oid
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get Operator Activity Log
      tags:
      - Operators
  /cluster/{id}/operator/{oid}/reinstall:
    post:
      consumes:
      - application/json
      description: ReInstalling Operator
      parameters:
      - description: Cluster Id
        in: path
        name: id
        required: true
        type: integer
      - description: Operator Id
        in: path
        name: oid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: ReInstall Operator
      tags:
      - Operators
  /cluster/{id}/operator/install:
    post:
      consumes:
      - application/json
      description: Install Operator
      parameters:
      - description: Operator Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.OperatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.Operator'
      security:
      - ApiKeyAuth: []
      summary: Install Operator
      tags:
      - Operators
  /cluster/{id}/operator/status:
    get:
      consumes:
      - application/json
      description: Operator Status
      parameters:
      - description: Cluster Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Operator Status
      tags:
      - Operators
  /cluster/{id}/operators:
    get:
      consumes:
      - application/json
      description: Get Organization Operators
      parameters:
      - description: Cluster Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doc.OperatorRequest'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get Organization Operators
      tags:
      - Operators
...
  /operator:
    get:
      consumes:
      - application/json
      description: Get operator
      parameters:
      - description: Package Name
        in: query
        name: packageName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.Operator'
      security:
      - ApiKeyAuth: []
      summary: Get operator
      tags:
      - Operators
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"